restructure api.views
