+++ backend/api/views/__init__.py
@@ -1,0 +1,4 @@
+from .get_nearby_or_create_factories import get_nearby_or_create_factories
+from .update_factory_attribute import update_factory_attribute
+from .post_image import post_image
+from .post_factory_image import post_factory_image
