+++ backend/api/views/update_factory_attribute.py
@@ -1,0 +1,7 @@
+from rest_framework.decorators import api_view
+
+
+@api_view(['PUT'])
+def update_factory_attribute(request, factory_id):
+    # TODO
+    pass
